feat: Rename files for associated output request/response on api
