+++ api/src/models/api/stardust/IAssociationsRequestBody.ts
@@ -1,6 +1,6 @@
 import { IBech32AddressDetails } from "./IBech32AddressDetails";
 
-export interface IAssociatedOutputsRequestBody {
+export interface IAssociationsRequestBody {
     /**
      * The address details of the address to get the associated outputs for.
      */
